refactor: useUserIdInput 훅 common으로 이동 #293
- 로그인, 회원가입 모두 공통으로 사용되어서 공통으로 이동
