+++ frontend/__tests__/useUserIdInput.test.ts
@@ -1,9 +1,9 @@
 import { renderHook, act } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
 
-import { ERROR_MESSAGE } from '../src/pages/signup/constants/errorMessage';
 import { ID } from '../src/common/constants/id';
-import useUserIdInput from '../src/pages/signup/hooks/useUserIdInput';
+import useUserIdInput from '../src/common/hooks/useUserIdInput';
+import { ERROR_MESSAGE } from '../src/pages/signup/constants/errorMessage';
 
 const createFakeChangeEvent = (
   value: string,
