+++ frontend/src/pages/signup/components/SignupForm/SignupForm.tsx
@@ -6,11 +6,11 @@ import styled from '@emotion/styled';
 import Button from '../../../../common/components/Button/Button';
 import useFormattedPhoneNumber from '../../../../common/hooks/useFormattedPhoneNumber';
 import useNameInput from '../../../../common/hooks/useNameInput';
+import useUserIdInput from '../../../../common/hooks/useUserIdInput';
 import { getPhoneNumberErrorMessage } from '../../../../common/utils/phoneNumberValidator';
 import { postSignup } from '../../apis/postSignup';
 import usePasswordWithConfirmInput from '../../hooks/usePasswordWithConfirmInput';
 import useUserIdDuplicateCheck from '../../hooks/useUserIdDuplicateCheck';
-import useUserIdInput from '../../hooks/useUserIdInput';
 import useVerificationCodeConfirm from '../../hooks/useVerificationCodeConfirm';
 import useVerificationCodeInput from '../../hooks/useVerificationCodeInput';
 import useVerificationCodeRequest from '../../hooks/useVerificationCodeRequest';
